refactor(data-dict): split extract_all.py into extract/ package
