+++ src/prenatalppkt/scripts/data_dict/extract/__init__.py
@@ -1,0 +1,1 @@
+"""Extract subpackage: walk Observer JSON + EVMS HL7 corpora into one CSV."""
